+++ react/hooks/useOutsideClick.js
@@ -1,13 +1,15 @@
-import { useLayoutEffect } from 'react'
+import { useLayoutEffect, useCallback } from 'react'
 
 const useOutsideClick = (ref, handler, activeWhen) => {
-  const handle = e =>
-    ref && ref.current && !ref.current.contains(e.target) && handler(e)
+  const handle = useCallback(
+    e => ref && ref.current && !ref.current.contains(e.target) && handler(e),
+    [handler, ref]
+  )
 
   useLayoutEffect(() => {
     activeWhen && document.addEventListener('mousedown', handle)
     return () => document.removeEventListener('mousedown', handle)
-  }, [activeWhen])
+  }, [activeWhen, handle])
 }
 
 export default useOutsideClick