+++ react/index.js
@@ -4,10 +4,8 @@ import SearchResultContainer from './components/SearchResultContainer'
 import { SORT_OPTIONS } from './OrderBy'
 import LocalQuery from './components/LocalQuery'
 import { LAYOUT_MODE } from './components/LayoutModeSwitcher'
-import GapPaddingTypes, {
-  getGapPaddingNames,
-  getGapPaddingValues,
-} from './constants/paddingEnum'
+import QueryContext from './components/QueryContext'
+import SettingsContext from './components/SettingsContext'
 
 const PAGINATION_TYPES = ['show-more', 'infinite-scroll']
 const DEFAULT_MAX_ITEMS_PER_PAGE = 10
@@ -19,8 +17,6 @@ const DEFAULT_MAX_ITEMS_PER_PAGE = 10
 export default class SearchResultQueryLoader extends Component {
   static defaultProps = {
     orderBy: SORT_OPTIONS[0].value,
-    gap: GapPaddingTypes.SMALL.value,
-    rest: '',
     showCategoryPanel: false,
     quantityOfItemsPerRow: 4,
   }
@@ -35,16 +31,40 @@ export default class SearchResultQueryLoader extends Component {
   }
 
   render() {
-    const { querySchema } = this.props
+    const {
+      querySchema,
+      hiddenFacets,
+      showFacetsQuantity,
+      pagination,
+      mobileLayout,
+    } = this.props
+
+    const settings = {
+      hiddenFacets,
+      showFacetsQuantity,
+      pagination,
+      mobileLayout,
+    }
+
     return !this.props.searchQuery ||
       (querySchema && querySchema.enableCustomQuery) ? (
       <LocalQuery
         {...this.props}
         {...querySchema}
-        render={props => <SearchResultContainer {...props} />}
+        render={props => (
+          <QueryContext.Provider value={props.searchQuery.variables}>
+            <SettingsContext.Provider value={settings}>
+              <SearchResultContainer {...props} />
+            </SettingsContext.Provider>
+          </QueryContext.Provider>
+        )}
       />
     ) : (
-      <SearchResultContainer {...this.props} />
+      <QueryContext.Provider value={this.props.searchQuery.variables}>
+        <SettingsContext.Provider value={settings}>
+          <SearchResultContainer {...this.props} />
+        </SettingsContext.Provider>
+      </QueryContext.Provider>
     )
   }
 }
@@ -53,12 +73,12 @@ SearchResultQueryLoader.getSchema = props => {
   const querySchema = !props.searchQuery
     ? {
         querySchema: {
-          title: 'editor.search-result.query',
-          description: 'editor.search-result.query.description',
+          title: 'admin/editor.search-result.query',
+          description: 'admin/editor.search-result.query.description',
           type: 'object',
           properties: {
             maxItemsPerPage: {
-              title: 'editor.search-result.query.maxItemsPerPage',
+              title: 'admin/editor.search-result.query.maxItemsPerPage',
               type: 'number',
               default: DEFAULT_MAX_ITEMS_PER_PAGE,
             },
@@ -68,10 +88,6 @@ SearchResultQueryLoader.getSchema = props => {
             },
             mapField: {
               title: 'Map',
-              type: 'string',
-            },
-            restField: {
-              title: 'Other Query Strings',
               type: 'string',
             },
             orderByField: {
@@ -87,18 +103,18 @@ SearchResultQueryLoader.getSchema = props => {
     : {}
 
   return {
-    title: 'editor.search-result.title',
-    description: 'editor.search-result.description',
+    title: 'admin/editor.search-result.title',
+    description: 'admin/editor.search-result.description',
     type: 'object',
     properties: {
       ...querySchema,
       mobileLayout: {
-        title: 'editor.search-result.mobileLayout',
+        title: 'admin/editor.search-result.mobileLayout',
         type: 'object',
         isLayout: true,
         properties: {
           mode1: {
-            title: 'editor.search-result.mobileLayout.mode1',
+            title: 'admin/editor.search-result.mobileLayout.mode1',
             type: 'string',
             default: LAYOUT_MODE[0].value,
             enum: LAYOUT_MODE.map(opt => opt.value),
@@ -106,7 +122,7 @@ SearchResultQueryLoader.getSchema = props => {
             isLayout: true,
           },
           mode2: {
-            title: 'editor.search-result.mobileLayout.mode2',
+            title: 'admin/editor.search-result.mobileLayout.mode2',
             type: 'string',
             default: LAYOUT_MODE[1].value,
             enum: LAYOUT_MODE.map(opt => opt.value),
@@ -116,33 +132,34 @@ SearchResultQueryLoader.getSchema = props => {
         },
       },
       hiddenFacets: {
-        title: 'editor.search-result.hiddenFacets',
+        title: 'admin/editor.search-result.hiddenFacets',
         type: 'object',
         isLayout: true,
         properties: {
           brands: {
-            title: 'editor.search-result.hiddenFacets.brands',
+            title: 'admin/editor.search-result.hiddenFacets.brands',
             type: 'boolean',
             isLayout: true,
           },
           categories: {
-            title: 'editor.search-result.hiddenFacets.categories',
+            title: 'admin/editor.search-result.hiddenFacets.categories',
             type: 'boolean',
             isLayout: true,
           },
           priceRange: {
-            title: 'editor.search-result.hiddenFacets.priceRange',
+            title: 'admin/editor.search-result.hiddenFacets.priceRange',
             type: 'boolean',
             isLayout: true,
           },
           specificationFilters: {
-            title: 'editor.search-result.hiddenFacets.specificationFilters',
+            title:
+              'admin/editor.search-result.hiddenFacets.specificationFilters',
             type: 'object',
             isLayout: true,
             properties: {
               hideAll: {
                 title:
-                  'editor.search-result.hiddenFacets.specificationFilters.hideAll',
+                  'admin/editor.search-result.hiddenFacets.specificationFilters.hideAll',
                 type: 'boolean',
                 isLayout: true,
               },
@@ -151,13 +168,13 @@ SearchResultQueryLoader.getSchema = props => {
                 isLayout: true,
                 items: {
                   title:
-                    'editor.search-result.hiddenFacets.specificationFilters.hiddenFilter',
+                    'admin/editor.search-result.hiddenFacets.specificationFilters.hiddenFilter',
                   type: 'object',
                   isLayout: true,
                   properties: {
                     name: {
                       title:
-                        'editor.search-result.hiddenFacets.specificationFilters.hiddenFilter.name',
+                        'admin/editor.search-result.hiddenFacets.specificationFilters.hiddenFilter.name',
                       type: 'string',
                       isLayout: true,
                     },
@@ -168,22 +185,14 @@ SearchResultQueryLoader.getSchema = props => {
           },
         },
       },
-      gap: {
-        title: 'editor.search-result.gap.title',
-        type: 'string',
-        enum: getGapPaddingValues(),
-        enumNames: getGapPaddingNames(),
-        default: GapPaddingTypes.SMALL.value,
-        isLayout: true,
-      },
       pagination: {
         type: 'string',
-        title: 'editor.search-result.pagination.title',
+        title: 'admin/editor.search-result.pagination.title',
         default: 'infinity-scroll',
         enum: PAGINATION_TYPES,
         enumNames: [
-          'editor.search-result.pagination.show-more',
-          'editor.search-result.pagination.infinite-scroll',
+          'admin/editor.search-result.pagination.show-more',
+          'admin/editor.search-result.pagination.infinite-scroll',
         ],
       },
       showCategoryPanel: {
@@ -204,7 +213,12 @@ SearchResultQueryLoader.getSchema = props => {
           },
         },
         isLayout: true,
-      }
+      },
+      showFacetQuantity: {
+        type: 'boolean',
+        title: 'admin/editor.search-result.facet-quantity',
+        default: false,
+      },
     },
   }
 }